1、meta

+++ data-model-manager/src/main/java/org/welisdoon/metadata/prototype/define/MetaObject.java
@@ -98,6 +98,8 @@ public class MetaObject extends MetaPrototype implements ITypeEntity<ObjectMetaT
 
     @Override
     public int remove() {
+        if (getState() == LifeState.Delete)
+            return 0;
         int update = getChildren().stream().map(MetaObject::remove).reduce(0, Integer::sum);
         super.remove();
         if (getConstructId() != null) {
@@ -161,6 +163,8 @@ public class MetaObject extends MetaPrototype implements ITypeEntity<ObjectMetaT
 
         @Override
         public int remove() {
+            if (getState() == LifeState.Delete)
+                return 0;
             super.remove();
             int update = MetaUtils.getInstance().getMetaObjectDao().delete(this.getId());
             if (getParentId() != null) {
